refactor: move core.py and new files to core/ folder
This makes importing from tinytuya.core continue to work.
