+++ tinytuya/core/crypto_helper.py
@@ -1,7 +1,6 @@
 # TinyTuya Module
 # -*- coding: utf-8 -*-
 
-# Modules
 from __future__ import print_function  # python 2.7 support
 import base64
 import logging
