+++ tinytuya/core/exceptions.py
